requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
colour==0.1.5
cycler==0.11.0
easyocr==1.6.0
fonttools==4.37.1
idna==3.3
imageio==2.21.1
kiwisolver==1.4.4
matplotlib==3.5.3
networkx==2.8.6
ninja==1.10.2.3
numpy==1.23.2
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.2.0
pyclipper==1.3.0.post3
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
Shapely==1.8.4
six==1.16.0
tifffile==2022.8.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
